python/pyproject.toml
[project]
name = "mix-cookbooks-python"
version = "0.1.0"
description = "Python examples for Mix SDK"
requires-python = ">=3.10"
dependencies = [
    "mix-python-sdk>=0.5.6",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []
